pyproject.toml
[tool.poetry]
name = "label_selection_layer"
version = "0.1.0"
description = ""
authors = ["ssatsuki <ssatsuki040508@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
torch = "^1.12.0"
pytorch-lightning = "^1.6.5"
PyYAML = "^6.0"
crowd-kit = "^1.0.0"
wandb = "^0.15.2"
numpy = "^1.24.3"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"

[tool.poetry.dev-dependencies]
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
